Rename repo to repository and improve config reading and loading
