+++ cmd/deepsource/main.go
@@ -1,10 +1,12 @@
 package main
 
 import (
+	"fmt"
 	"log"
 	"os"
 
 	"github.com/deepsourcelabs/cli/command"
+	"github.com/deepsourcelabs/cli/config"
 	v "github.com/deepsourcelabs/cli/version"
 	"github.com/getsentry/sentry-go"
 	"github.com/pterm/pterm"
@@ -25,11 +27,19 @@ func main() {
 
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
 
+	// Init sentry
 	err := sentry.Init(sentry.ClientOptions{
 		Dsn: SentryDSN,
 	})
 	if err != nil {
 		log.Println("Could not load sentry.")
+	}
+
+	// Load the config data
+	err = loadConfig()
+	if err != nil {
+		pterm.Error.Println(err)
+		os.Exit(1)
 	}
 
 	v.SetBuildInfo(version, Date, "", "")
@@ -39,4 +49,19 @@ func main() {
 		pterm.Error.Println(err)
 		os.Exit(1)
 	}
+}
+
+// Loads the config data
+func loadConfig() error {
+	// Initialize and load the DeepSource CLI config file
+	if err := config.InitConfig(); err != nil {
+		return fmt.Errorf("Error reading config file: %v", err)
+	}
+
+	// Check if token expired
+	if config.Cfg.Token != "" && config.Cfg.IsExpired() {
+		pterm.Info.Println("The authentication has expired. Please refresh the token using `deepsource auth refresh`")
+	}
+	// TODO: Automatically run `deepsource auth refresh` here
+	return nil
 }
