+++ command/repo/status/status.go
@@ -25,7 +25,7 @@ func NewCmdRepoStatus() *cobra.Command {
 
 	opts := RepoStatusOptions{
 		RepoArg:      "",
-		TokenExpired: config.TokenExpired,
+		TokenExpired: config.Cfg.IsExpired(),
 		Owner:        "",
 		RepoName:     "",
 		VCSProvider:  "",
