+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/deepsourcelabs/cli
 
-go 1.12
+go 1.16
 
 require (
 	github.com/AlecAivazis/survey/v2 v2.2.12
